// custom_typings/enums.ts
export enum HorizontalAlignValues {
    LEFT = "left",
    CENTER = "center",
    RIGHT = "right"
}

export enum VerticalAlignValues {
    TOP = "top",
    MIDDLE = "middle",
    BOTTOM = "bottom"
}

export enum FormattersValues {
    PLAINTEXT = "plaintext",
    TEXTAREA = "textarea",
    HTML = "html",
    MONEY = "money",
    IMAGE = "image",
    LINK = "link",
    DATETIME = "datetime",
    DATETIME_DIFF = "datetimediff",
    TICK_CROSS = "tickCross",
    COLOR = "color",
    START = "star",
    TRAFFIC = "traffic",
    PROGRESS = "progress",
    LOOKUP = "lookup",
    BUTTON_TICK = "buttonTick",
    BUTTON_CROSS = "buttonCross",
    ROW_NUM = "rownum",
    HANDLE = "handle"
}

export enum SortersValues {
    STRING = "string",
    NUMBER = "number",
    ALPHANUM = "alphanum",
    BOOLEAN = "boolean",
    EXISTS = "exists",
    DATE = "date",
    TIME = "time",
    DATETIME = "datetime",
    ARRAY = "array"
}

export enum BrokeragePortfolioTypes {
    BROKER_ACCOUNT = "brokerAccount",
    MODEL_PORTFOLIO = "modelPortfolio"
}

export enum EditableTableColumns {
    NAME = "name",
    WEIGHT = "weight",
    BRIEFCASE = "briefcase",
    GROUP_NAME = "groupName"
}

export enum SidebarMenuElementsTypes {
    BROKER_ACCOUNT = "brokerAccount",
    MODEL_PORTFOLIO = "modelPortfolio"
}